google-play-scraper/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
google-play-scraper==1.2.4
openai==1.12.0
python-dotenv==1.0.1
pydantic==2.6.1 